chore: upgrade to veramo 1.0.1
