+++ .vscode/launch.json
@@ -11,7 +11,7 @@
       "skipFiles": [
         "<node_internals>/**"
       ],
-      "program": "${workspaceFolder}/node_modules/daf-cli/build/cli.js",
+      "program": "${workspaceFolder}/node_modules/@veramo/cli/build/cli.js",
       "args": ["server"]
     }
   ]
