+++ src/index.ts
@@ -7,6 +7,6 @@ export { MyAgentPlugin } from './agent/my-plugin'
 export { KeyManagementSystem } from './key-manager/key-management-system'
 export { KeyStore } from './key-manager/key-store'
 export { SecretBox } from './key-manager/secret-box'
-export { IdentityProvider } from './identity-manager/identity-provider'
-export { IdentityStore } from './identity-manager/identity-store'
+export { IdentifierProvider } from './did-manager/identifier-provider'
+export { DIDStore } from './did-manager/did-store'
 export * from './types/IMyAgentPlugin'
