+++ src/key-manager/secret-box.ts
@@ -1,4 +1,4 @@
-import { AbstractSecretBox } from 'daf-key-manager'
+import { AbstractSecretBox } from '@veramo/key-manager'
 
 /**
  * @alpha
